refactor: rename agent_execution to execution and enhance module exports - Rename agent_execution > execution directory - Update all import paths across codebase to use new module structure
