--- libs/aion-server/src/aion/server/agent/agent_execution/__init__.py
@@ -1,5 +1,0 @@
-from .request_executor import AionAgentRequestExecutor
-
-__all__ = [
-    "AionAgentRequestExecutor",
-]
